Allow the solver to use a lookup table
Splits solver functionality into a new class.

Each turn, the solver will check a lookup table in `./data/lookup.json`.
If it find the current game state, it will use that instead of searching
the tree. If it searches the tree, the best move will be added to the
lookup table and saved to `./data/new_lookup.json`

Signed-off-by: Kurtis Charnock <kurtis@charno.cc>